Create Category

+++ src/main/java/com/meloDeveloper/spring_boot_learning/entities/Order.java
@@ -22,7 +22,7 @@ public class Order implements Serializable{
 	
 	@Id
 	@GeneratedValue(strategy=GenerationType.IDENTITY)
-	private Integer id;
+	private Long id;
 	@JsonFormat(shape=JsonFormat.Shape.STRING,pattern = "yyyy-dd-MM'T'hh:mm:ss'Z'",timezone = "GMT")
 	private Instant moment;
 	private int orderStatus;
@@ -33,18 +33,18 @@ public class Order implements Serializable{
 	public Order() {
 	}
 	
-	public Order(Integer id, Instant moment, OrderStatus orderStatus, User client) {
+	public Order(Long id, Instant moment, OrderStatus orderStatus, User client) {
 		this.id = id;
 		this.moment = moment;
 		setOrderStatus(orderStatus);
 		this.client = client;
 	}
 
-	public Integer getId() {
+	public Long getId() {
 		return id;
 	}
 
-	public void setId(Integer id) {
+	public void setId(Long id) {
 		this.id = id;
 	}
